refactor(scrollable-video): replace `onCanPlayThrough` by `onLoadedData`
The previous commit removes `autoPlay` props.
This patch adds it back.
`onLoadedData` event is triggered on iOS only if `autoPlay` is set to
`true`. For other devices rather than iOS, we actually don't need
`autoPlay` here since we use JavaScript to manipulate video
playing/pausing.

In this patch, we also replace `onCanPlayThrough` by `onLoadedData`.
This change is for better user experience.
`onCanPlayThrough` is triggered when the video is fully loaded, and
`onLoadedData` is triggered when the first frame of video is loaded.

We don't want users to wait for video fully downloading. We want them to
see the video frames as soon as possible.