fix(scrollable-image): progress jumps due to `100vh` changes
This patch fixes a bug where the progress of web page jumps due to
`100vh` pixels changes on few browsers (e.g. Facebook in-app browser).

In this change, we keep the largest pixel value for 100vh and uses it
as the height of every images as well as their containers. Since the
dimention of images fixes all the time, the height of
the wrapper won't change when `100vh` pixel value update.